+++ src/main/java/com/stosik/parking/reservation/domain/evaluator/RateEvaluatorStrategy.java
@@ -2,7 +2,9 @@ package com.stosik.parking.reservation.domain.evaluator;
 
 import com.stosik.parking.reservation.domain.model.Reservation;
 
+import java.math.BigDecimal;
+
 interface RateEvaluatorStrategy
 {
-    double calculateReservationCost(Reservation reservation);
+    BigDecimal calculateReservationCost(Reservation reservation);
 }
